Initial server firmware
This is just a basic hello world at this point.   It initializes the Wifi into access
point mode and then listens as both an HTTP server as well as on a UDP port.
It displays some basic information on the screen and then just serves
on http://192.168.4.1/ and on UDP port 8888.