More efficient update and first pass on Station data model
I've removed the delays in the prior server instance since we
may have been seeing some packetloss and delays due to that.
Also more efficient updates of the screen with bounding boxes
that are tight around the information that is being changed.

Added a first pass on the station data model.