First pass on showing results
After talking with Aidan and discussing the use cases I've created this class that shows a descritized view of the data.  This view current draws from the top to the bottom and in a single color but captures the gist of what we were talking about which mimics a sound board with specific steps that show the current sound level and a bar that shows the trailing maximum value.

Documentation and better code is forthcoming.