Adding an image to capture the new exception requirement.
In order to allow exceptions to be used within our software, which I use in the new Graphing component, you have to enable exceptions as a build parameter within the tools menu.  Change 'Exceptions' to "Enabled" otherwise you will see an error about exceptions not being enabled and having to use the -fexceptions flag.