Proper update of Graph
This change ensures that the graph is updated 33 times a second and that any
data not received in this period will be interpreted as having a default (0) value.
I've also updated the mechanism to send packets to allow a normal distribution
sent over time.  To use it:

C:\Users\mchug\Documents\GitHub\AMS>python sendExampleSequence.py 5 10