Moved Server and added debugging
Based on a conversation with Aidan we agreed that the server instance would benefit from outputting diagnostic information which could be used after-the-fact to understand what was happening.  This will be useful if/when odd things occur with the system so that we can sit together and view the output and understand what happened.

I've also moved the server firmware files to a simpler directory structure.