Sample the audio sensor for the waveform
In this change I've made the node software look at the sensor across a range of
time (currently 5 milliseconds) and extract from that the min, max and sample count sizes.  The difference in raw min and max values gives us the magnitude of the relative waveform and thus a sense of the loudness of the noise within the environment.